backend/pyproject.toml
[project]
name = "PasswordManager"
version = "0.1.0"
description = "Password Manager web application"
authors = [{ name = "mrrobot" }]
requires-python = ">=3.8"

dependencies = [
    "django>=4.2",
    "django-cors-headers>=4.4.0",
    "django-ninja>=1.4.1",
    "django-stubs>=5.1.3",
    "django-unfold>=0.46.0",
    "dj-database-url>=2.1.0",
    "openai>=1.84.0",
    "pillow>=10.4.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "PyJWT>=2.8.0",
    "ruff>=0.11.8",
    "uvicorn>=0.33.0",
    "whitenoise>=6.6.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "netaddr>=0.10.1",
]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "ty>=0.0.1a5",
]
